move dupenv to common
