+++ common/dupenv.c
