+++ common/dupenv.h
